+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [NEXT]
 
+### Fixed
+- Fixed missing claim validation, now also validates the attestation claims `nbf` and `iat` (#1066).
+
 ## [4.0.1] - 2026-07-09
 
 ### Fixed